+++ edu/tomanova/splines/exact/tymoshenko/SquareSolver.py
