+++ edu/tomanova/splines/resources/square/exactSolutionByTymoshenko.png
